Rename no-comment-textnodes to jsx-no-comment-textnodes
This rule is JSX-specific, not React-specific. Changing its name makes
this clearer. I've kept the old one around as deprecated. We can remove
it at the next major version bump.

Addresses #668 and #686.
