+++ index.js
@@ -8,6 +8,7 @@ var rules = {
   'wrap-multilines': require('./lib/rules/wrap-multilines'),
   'self-closing-comp': require('./lib/rules/self-closing-comp'),
   'no-comment-textnodes': require('./lib/rules/no-comment-textnodes'),
+  'jsx-no-comment-textnodes': require('./lib/rules/jsx-no-comment-textnodes'),
   'no-danger': require('./lib/rules/no-danger'),
   'no-set-state': require('./lib/rules/no-set-state'),
   'no-is-mounted': require('./lib/rules/no-is-mounted'),
