+++ lib/rules/no-comment-textnodes.js
@@ -1,6 +1,7 @@
 /**
  * @fileoverview Comments inside children section of tag should be placed inside braces.
  * @author Ben Vinegar
+ * @deprecated
  */
 'use strict';
 
@@ -8,27 +9,24 @@
 // Rule Definition
 // ------------------------------------------------------------------------------
 
+var util = require('util');
+var jsxNoCommentTextnodes = require('./jsx-no-comment-textnodes');
+var isWarnedForDeprecation = false;
+
 module.exports = function(context) {
-  function reportLiteralNode(node) {
-    context.report(node, 'Comments inside children section of tag should be placed inside braces');
-  }
-
-  // --------------------------------------------------------------------------
-  // Public
-  // --------------------------------------------------------------------------
-
-  return {
-    Literal: function(node) {
-      if (/^\s*\/(\/|\*)/m.test(node.value)) {
-        // inside component, e.g. <div>literal</div>
-        if (node.parent.type !== 'JSXAttribute' &&
-            node.parent.type !== 'JSXExpressionContainer' &&
-            node.parent.type.indexOf('JSX') !== -1) {
-          reportLiteralNode(node);
-        }
+  return util._extend(jsxNoCommentTextnodes(context), {
+    Program: function() {
+      if (isWarnedForDeprecation || /\=-(f|-format)=/.test(process.argv.join('='))) {
+        return;
       }
+
+      /* eslint-disable no-console */
+      console.log('The react/no-comment-textnodes rule is deprecated. Please ' +
+                  'use the react/jsx-no-comment-textnodes rule instead.');
+      /* eslint-enable no-console */
+      isWarnedForDeprecation = true;
     }
-  };
+  });
 };
 
 module.exports.schema = [{
